// src/app/core/models/transportista.ts
export class Transportista {
    codigo: string; //este es el campo que funciona de Id
    nombre: string;
    direccion: string;
    provincia: string;
    cuit: string;
    telefonoEmpresa: string;
    casillaCorreo: string;
    telefonoContacto: string;
    localidad: string;
    codigoNombre: string;
    idDescripcion:string;
}